requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
httpx
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
python-dotenv
pydantic-settings
python-multipart
pytest-cov
langchain[all]
langchain-community
pymupdf
langchain-ollama
transformers
tokenizers
pytest-mock
aiosqlite
asgi_lifespan